pyproject.toml
[project]
name = "jett"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  # Web framework
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  # HTTP client for service-to-service calls
  "httpx>=0.25.0",
  # Pydantic for data validation
  "pydantic>=2.5.0",
  "pydantic-settings>=2.1.0",
  # PDF processing
  "pypdf>=4.0.0",
  "pdf2image>=1.16.0",
  # Image processing
  "pillow>=10.1.0",
  # File uploads
  "python-multipart>=0.0.6",
  # Logging
  "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "pytest-asyncio>=0.21.0",
  "black>=23.0.0",
  "ruff>=0.1.0",
  "mypy>=1.6.0",
]
ppocrv4-server = [
  "paddlepaddle>=3.0.0",
  "paddleocr>=2.8",
  "numpy<2.0",
  "opencv-python-headless>=4.8",
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "python-multipart>=0.0.6",
  "pillow>=10.1.0",
  "pydantic>=2.5.0",
]
ppocrv4-mobile = [
  "paddlepaddle>=3.0.0",
  "paddleocr>=2.8",
  "numpy<2.0",
  "opencv-python-headless>=4.8",
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "python-multipart>=0.0.6",
  "pillow>=10.1.0",
  "pydantic>=2.5.0",
]
easyocr = [
  "easyocr>=1.7",
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "python-multipart>=0.0.6",
  "pillow>=10.1.0",
  "pydantic>=2.5.0",
  "numpy>=1.26",
  "opencv-python-headless>=4.8",
]
rapidocr = [
  "rapidocr-onnxruntime>=1.3",
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "python-multipart>=0.0.6",
  "pillow>=10.1.0",
  "pydantic>=2.5.0",
  "numpy>=1.26",
  "opencv-python-headless>=4.8",
]

[tool.uv]
default-packages = []
